Refactored Pandera schema error tranformation into separate module.
